+++ plug-api/syscalls/index.ts
@@ -31,9 +31,8 @@ export function queryLuaObjects<T>(
   tag: string,
   query: LuaCollectionQuery,
   scopedVariables?: Record<string, any>,
-  ttlSecs?: number,
 ): Promise<ObjectValue<T>[]> {
-  return syscall("index.queryLuaObjects", tag, query, scopedVariables, ttlSecs);
+  return syscall("index.queryLuaObjects", tag, query, scopedVariables);
 }
 
 /**
@@ -56,4 +55,16 @@ export function getObjectByRef<T>(
  */
 export function ensureFullIndex(): Promise<void> {
   return syscall("index.ensureFullIndex");
+}
+
+export function reindexSpace(): Promise<void> {
+  return syscall("index.reindexSpace");
+}
+
+export function deleteObject(
+  page: string,
+  tag: string,
+  ref: string,
+): Promise<void> {
+  return syscall("index.deleteObject", page, tag, ref);
 }