+++ plugs/index/command.ts
@@ -1,8 +1,7 @@
-import { editor } from "@silverbulletmd/silverbullet/syscalls";
-import { reindexSpace } from "./queue.ts";
+import { editor, index } from "@silverbulletmd/silverbullet/syscalls";
 
 export async function reindexCommand() {
   await editor.flashNotification("Performing full page reindex...");
-  await reindexSpace();
+  await index.reindexSpace();
   await editor.flashNotification("Done with page index!");
 }