+++ plugs/index/plug_api.ts
@@ -1,7 +1,7 @@
 import { system } from "@silverbulletmd/silverbullet/syscalls";
 import type { ObjectValue } from "../../plug-api/types/index.ts";
 
-import type { KV, KvQuery } from "../../plug-api/types/datastore.ts";
+import type { KV } from "../../plug-api/types/datastore.ts";
 
 export function indexObjects<T>(
   page: string,
@@ -12,12 +12,6 @@ export function indexObjects<T>(
 
 export function batchSet(page: string, kvs: KV[]): Promise<void> {
   return system.invokeFunction("index.batchSet", page, kvs);
-}
-
-export function query(
-  query: KvQuery,
-): Promise<KV[]> {
-  return system.invokeFunction("index.query", query);
 }
 
 export function getObjectByRef<T>(